requirements.txt
newspaper3k
joblib
tqdm
pandas
gunicorn
Flask